Adds: Website/Store ID filtering on product sync

+++ src/Support/MagentoProducts.php
@@ -105,7 +105,9 @@ class MagentoProducts extends PaginatableMagentoService
     {
         $product = MagentoProduct::where('sku', $sku)->first();
 
-        $product->delete();
+        if ($product) {
+            $product->delete();
+        }
 
         return $this;
     }
